@@ -32,56 +32,61 @@
     "migrate:make": "node ./bin/db.js migrate:make",
     "migrate:teardown": "node ./bin/db.js migrate:teardown",
     "migrate:teardown:test": "NODE_ENV=testing node ./bin/db.js migrate:teardown",
-    "migrate:reset": "npm run migrate:teardown && npm run migrate:latest",
-    "migrate:reset:test": "npm run migrate:teardown:test && npm run migrate:latest:test",
+    "db:reset": "npm run migrate:teardown && npm run migrate:latest",
+    "db:reset:test": "npm run migrate:teardown:test && npm run migrate:latest:test",
     "seed:make": "knex seed:make",
     "seed:run": "knex seed:run",
     "seed:run:test": "NODE_ENV=testing knex seed:run",
-    "dev": "tsc -w & nodemon ./build/src/index.js",
+    "dev": "nodemon",
+    "copy:assets": "cp -R ./assets ./build/assets",
+    "copy:fonts": "cp -R ./src/services/pdf/fonts ./build/src/services/pdf/fonts",
+    "copy": "npm run copy:assets && npm run copy:fonts",
     "start": "node ./build/src/index.js",
     "watch": "tsc -w",
-    "build": "tsc",
+    "build": "tsc && npm run copy",
     "heroku-postbuild": "npm run build && npm run migrate:latest"
   },
   "dependencies": {
     "@types/axios": "^0.14.0",
-    "@types/bcrypt": "^2.0.0",
-    "@types/hapi": "^17.0.19",
-    "@types/hapi-auth-jwt2": "^8.0.1",
-    "@types/jest": "^23.3.1",
-    "@types/jsonwebtoken": "^7.2.8",
-    "@types/knex": "^0.14.23",
-    "@types/pdfmake": "^0.1.1",
-    "@types/ramda": "^0.25.37",
+    "@types/bcrypt": "^3.0.0",
+    "@types/hapi": "^17.6.3",
+    "@types/hapi-auth-jwt2": "^8.0.2",
+    "@types/jest": "^23.3.9",
+    "@types/jsonwebtoken": "^8.3.0",
+    "@types/knex": "^0.15.0",
+    "@types/pdfmake": "^0.1.2",
+    "@types/ramda": "^0.25.41",
     "axios": "^0.18.0",
-    "bcrypt": "^3.0.0",
-    "boom": "^7.2.0",
-    "dotenv": "^6.0.0",
+    "bcrypt": "^3.0.2",
+    "boom": "^7.2.2",
+    "chalk": "^2.4.1",
+    "dotenv": "^6.1.0",
     "good": "^8.1.1",
     "good-console": "^7.1.0",
     "good-squeeze": "^5.1.0",
-    "hapi": "^17.5.4",
+    "hapi": "^17.7.0",
     "hapi-auth-bearer-token": "^6.1.0",
     "hapi-auth-jwt2": "^8.1.0",
-    "hapi-pino": "^4.0.4",
-    "joi": "^13.6.0",
+    "hapi-qs": "^2.0.1",
+    "joi": "^14.0.4",
     "jsonwebtoken": "^8.3.0",
     "knex": "^0.15.2",
     "minimist": "^1.2.0",
     "moment": "^2.22.2",
-    "pdfmake": "^0.1.38",
-    "pg": "^7.4.3",
-    "postmark": "^1.6.1",
-    "qrcode": "^1.2.2",
+    "pdfmake": "^0.1.39",
+    "pg": "^7.6.1",
+    "postmark": "^2.2.0",
+    "qrcode": "^1.3.2",
     "ramda": "^0.25.0",
-    "typescript": "^3.0.3"
+    "ts-node": "^7.0.1",
+    "typescript": "^3.1.6"
   },
   "devDependencies": {
     "axios-mock-adapter": "^1.15.0",
-    "factory-girl": "^5.0.2",
-    "jest": "^23.5.0",
-    "nodemon": "^1.18.4",
-    "ts-jest": "^23.1.4",
+    "factory-girl": "^5.0.3",
+    "jest": "^23.6.0",
+    "nodemon": "^1.18.7",
+    "ts-jest": "^23.10.4",
     "tslint-config-twine": "github:TwinePlatform/tslint-config-twine#2.1.0"
   }
 }
