@@ -4,7 +4,7 @@ exports.seed = (knex) =>
       { access_role_name: 'VISITOR' },
       { access_role_name: 'VOLUNTEER' },
       { access_role_name: 'VOLUNTEER_ADMIN' },
-      { access_role_name: 'ORG_ADMIN' },
+      { access_role_name: 'CB_ADMIN' },
       { access_role_name: 'FUNDING_BODY' },
       { access_role_name: 'TWINE_ADMIN' },
       { access_role_name: 'SYS_ADMIN' },
