@@ -1,45 +1,15 @@
-const api = require('../../src/api/v1/api.json');
-const perms = require('../../src/api/v1/permissions.json');
-const { map } = require('ramda');
+const perms = require('../seeds/permissions.seed.json');
 
 const restrictedScopes = {
-  ORG_ADMIN: [
+  CB_ADMIN: [
     "organisations_details-own:read",
     "organisations_feedback-own:write",
     "visit_activities-own:read",
-    "visit_logs-own:write",
+    "visit_logs-child:write",
+    "user_details-child:write",
+    "user_details-child:read",
   ]
 }
-
-const permissionsForRoles = Object.values(api.routes)
-  .reduce((acc, nestedRoutes) => {
-    const route = Object.values(nestedRoutes)
-      .reduce((accRouteValues, routeObject) =>
-        [...accRouteValues, ...Object.values(routeObject)]
-      ,[])
-      .map(({ scope, intendedFor, description }) => ({ scope, intendedFor, description }))
-    return [...acc, ...route]
-  }, [])
-  .reduce((acc, el) => {
-    el.intendedFor.forEach(x => {
-      if (el.scope.length > 0 && el.auth !== 'external') {
-        el.scope.forEach(y => acc[x].add(y))
-      }
-    })
-    return acc
-  }, {
-  TWINE_ADMIN: new Set(),
-  FUNDING_BODY: new Set(),
-  ORG_ADMIN: new Set(),
-  VOLUNTEER: new Set(),
-  VOLUNTEER_ADMIN: new Set(),
-});
-
-const allPermissions = Object.values(permissionsForRoles)
-  .reduce((acc, el) => {
-    el.forEach(x => acc.add(x))
-    return acc
-}, new Set());
 
 const rx = new RegExp('[-:]');
 
@@ -88,13 +58,3 @@ module.exports = {
   permissionRows: Array.from(perms.permissions).map(scopeToPermission),
   accessRolePermissionsRows,
 };
-
-
-if (require.main === module) {
-  console.log(
-    JSON.stringify({
-      'permissions': Array.from(allPermissions).sort(),
-      'permissionsForRoles': map((x) => Array.from(x).sort(), permissionsForRoles)
-    })
-  );
-}
