@@ -21,10 +21,49 @@ describe('GET /community-businesses', () => {
   });
 
   describe('GET /community-businesses', () => {
+    test('success:: user TWINE_ADMIN return list of all cbs', async () => {
+      const res = await server.inject({
+        method: 'GET',
+        url: '/v1/community-businesses',
+        credentials: {
+          scope: ['organisations_details-child:read'],
+          user: await Users.getOne(knex, { where: { name: 'Big Boss' } }),
+          organisation: await Organisations.getOne(knex, {
+            where: { name: 'Black Mesa Research' },
+          }),
+          role: RoleEnum.TWINE_ADMIN,
+        },
+      });
+
+      expect(res.statusCode).toBe(200);
+      expect((<any> res.result).result).toHaveLength(2);
+    });
+
+    test('success: admin code fields query returns subset of cbs', async () => {
+      const res = await server.inject({
+        method: 'GET',
+        url: '/v1/community-businesses?fields[]=adminCode&fields[]=name',
+        credentials: {
+          scope: ['organisations_details-child:read'],
+          user: await Users.getOne(knex, { where: { name: 'Big Boss' } }),
+          organisation: await Organisations.getOne(knex, {
+            where: { name: 'Black Mesa Research' },
+          }),
+          role: RoleEnum.TWINE_ADMIN,
+        },
+      });
+
+      expect(res.statusCode).toBe(200);
+      expect((<any> res.result).result).toEqual(expect.arrayContaining([
+        { adminCode: '10101', name: 'Aperture Science' },
+        { adminCode: '70007', name: 'Black Mesa Research' }])
+        );
+    });
+
     test('Fetching collection returns only list that user is authorised for', async () => {
       const res = await server.inject({
         method: 'GET',
-        url: '/api/v1/community-businesses',
+        url: '/v1/community-businesses',
         credentials: {
           scope: ['organisations_details-child:read'],
           user: await Users.getOne(knex, { where: { name: 'Gordon' } }),
@@ -34,18 +73,38 @@ describe('GET /community-businesses', () => {
         },
       });
 
-      expect(res.statusCode).toBe(404);
+      expect(res.statusCode).toBe(403);
     });
   });
 
   describe('GET /community-businesses/me', () => {
-    test('Returns CB that user is authenticated against', async () => {
+    test('Returns CB that CB_ADMIN is authenticated against', async () => {
       const res = await server.inject({
         method: 'GET',
         url: '/v1/community-businesses/me',
         credentials: {
           scope: ['organisations_details-own:read'],
           user: await Users.getOne(knex, { where: { name: 'Gordon' } }),
+          organisation: await Organisations.getOne(knex, {
+            where: { name: 'Black Mesa Research' },
+          }),
+        },
+      });
+
+      expect(res.statusCode).toBe(200);
+      expect(res.result).toEqual({
+        result: expect.objectContaining({ _360GivingId: 'GB-COH-9302' }),
+      });
+      expect(Object.keys((<any> res.result).result)).toHaveLength(15);
+    });
+
+    test('Returns CB that VOLUNTEER is authenticated against', async () => {
+      const res = await server.inject({
+        method: 'GET',
+        url: '/v1/community-businesses/me',
+        credentials: {
+          scope: ['organisations_details-parent:read'],
+          user: await Users.getOne(knex, { where: { name: 'Emma Emmerich' } }),
           organisation: await Organisations.getOne(knex, {
             where: { name: 'Black Mesa Research' },
           }),
