@@ -3,13 +3,11 @@
  */
 import * as Joi from 'joi';
 import { query } from '../users/schema';
+import { id } from '../schema/request';
 
 
-export { query } from '../schema/request';
+export { query, id, gender, since, until } from '../schema/request';
 export { response } from '../schema/response';
-
-export const since = Joi.date().iso().default(0);
-export const until = Joi.date().iso().default(() => Date.now(), 'Current date');
 
 export const visitActivitiesGetQuery = {
   ...query,
@@ -41,5 +39,4 @@ export const visitActivitiesPutPayload = {
   sunday: Joi.boolean(),
 };
 
-export const id = Joi.number().positive().required();
-export const meOrId = id.allow('me').required();
+export const meOrId = id.allow('me');
