@@ -19,7 +19,7 @@ import { Map } from '../../../types/internal';
 export type ApiRequestQuery = {
   fields: string[]
   sort: string
-  order: string
+  order: 'asc' | 'desc'
   offset: number
   limit: number
 };
@@ -33,7 +33,7 @@ export type ApiRequestBody = Dictionary<any>;
 export const query: Map<keyof ApiRequestQuery, Joi.Schema> = {
   fields: Joi.array().items(Joi.string().min(1).max(255)),
   sort: Joi.string(),
-  order: Joi.string(),
+  order: Joi.string().only('asc', 'desc'),
   offset: Joi.number().integer().min(0),
   limit: Joi.number().integer().min(1).max(100),
 };
@@ -47,3 +47,6 @@ export const gender =
 
 export const id =
   Joi.number().integer().positive();
+
+export const since = Joi.date().iso().default('1970-01-01T00:00:00.000Z');
+export const until = Joi.date().iso().default(() => Date.now(), 'Current date');
