@@ -2,11 +2,15 @@ import * as Hapi from 'hapi';
 import { Dictionary } from 'ramda';
 import { ApiRequestQuery, ApiRequestBody } from './schema/request';
 import { ApiResponse } from './schema/response';
-import { UserCredentials } from '../../auth/strategies/standard';
+import { UserCredentials as UC } from '../../auth/strategies/standard';
+import { RoleEnum } from '../../auth/types';
+import { GenderEnum, CommunityBusiness, User, CommonTimestamps, VolunteerLog } from '../../models';
+import { Omit } from '../../types/internal';
 
 
 declare module 'hapi' {
-  interface AuthCredentials extends UserCredentials {}
+  interface AuthCredentials extends UC {}
+  interface UserCredentials extends User {}
 }
 
 
@@ -18,6 +22,13 @@ export enum HttpMethodEnum {
   POST = 'POST',
   PUT = 'PUT',
   DELETE = 'DELETE',
+}
+
+export enum AppEnum {
+  TWINE_API = 'TWINE_API',
+  VISITOR = 'VISITOR_APP',
+  VOLUNTEER = 'VOLUNTEER_APP',
+  ADMIN = 'ADMIN_APP',
 }
 
 export type ApiRouteSpec = {
@@ -32,11 +43,38 @@ export type ApiRouteSpec = {
 };
 
 /*
+ * Prereq types
+ */
+export interface RequireSiblingPreReq extends Hapi.Request {
+  params: {
+    userId: string
+  };
+}
+
+/*
  * Request types
  */
-export interface OrganisationRequest extends Hapi.Request {
+export interface GetCommunityBusinessRequest extends Hapi.Request {
   params: {
     organisationId: string
+  };
+  query: ApiRequestQuery & {
+    [k: string]: any
+  };
+}
+
+export interface GetCommunityBusinessesRequest extends Hapi.Request {
+  query: ApiRequestQuery & {
+    [k: string]: any
+  };
+}
+
+export interface PutCommunityBusinesssRequest extends Hapi.Request {
+  payload:
+    Omit<CommunityBusiness, 'createdAt' | 'modifiedAt' | 'deletedAt' | 'id' | '_360GivingId'>;
+  pre: {
+    communityBusiness: CommunityBusiness
+    isChild?: boolean
   };
 }
 
@@ -55,14 +93,168 @@ export interface PostFeedbackRequest extends Hapi.Request {
 
 export interface LoginRequest extends Hapi.Request {
   payload: {
-    email: string,
-    password: string,
+    restrict?: RoleEnum | RoleEnum[]
+    type: 'cookie' | 'header'
+    email: string
+    password: string
   };
 }
 
 export interface EscalateRequest extends Hapi.Request {
   payload: {
     password: string
+  };
+}
+
+export interface GetVisitorsRequest extends Hapi.Request {
+  query: ApiRequestQuery & {
+    [k: string]: any
+    filter?: {
+      age?: [number, number]
+      gender?: GenderEnum
+      activity?: string
+      name?: string
+    }
+    visits: boolean
+  };
+}
+
+export interface GetVisitorRequest extends Hapi.Request {
+  params: {
+    userId: string
+  };
+  query: {
+    visits: string
+  };
+}
+
+export interface PutUserRequest extends Hapi.Request {
+  payload: Partial<Omit<User, 'id' | keyof CommonTimestamps | 'qrCode'>>;
+  params: {
+    userId: string
+  };
+}
+export interface GetAllVolunteersRequest extends Hapi.Request {
+  query: ApiRequestQuery & {
+    [k: string]: string
+  };
+}
+
+export interface DeleteUserRequest extends Hapi.Request {
+  params: {
+    userId: string
+  };
+}
+
+
+export interface GetMyVolunteerLogsRequest extends Hapi.Request {
+  query: ApiRequestQuery & {
+    since: string
+    until: string
+  };
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface PostMyVolunteerLogsRequest extends Hapi.Request {
+  payload: Pick<VolunteerLog, 'activity' | 'duration' | 'startedAt'> & {
+    userId?: number | 'me'
+  };
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface GetVolunteerLogRequest extends Hapi.Request {
+  query: { fields: (keyof VolunteerLog)[] };
+  params: { logId: string };
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface PutMyVolunteerLogRequest extends Hapi.Request {
+  params: { logId: string };
+  payload: Partial<Omit<VolunteerLog, 'id' | 'userId' | 'organisationId' | keyof CommonTimestamps>>;
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface GetVolunteerLogSummaryRequest extends Hapi.Request {
+  query: { since: string, until: string };
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface RegisterRequest extends Hapi.Request {
+  payload: {
+    organisationId: number
+    name: string
+    gender: GenderEnum
+    birthYear: number
+    email: string
+    phoneNumber: string
+    postCode: string
+    emailConsent: boolean
+    smsConsent: boolean
+  };
+}
+
+export interface VolunteerRegisterRequest extends Hapi.Request {
+  payload: RegisterRequest['payload'] & {
+    password: string
+    role: RoleEnum.VOLUNTEER | RoleEnum.VOLUNTEER_ADMIN
+    adminCode?: string
+  };
+}
+
+export interface GetMyVolunteerLogsAggregateRequest extends Hapi.Request {
+  query: ApiRequestQuery & Hapi.Util.Dictionary<string>;
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface SyncMyVolunteerLogsRequest extends Hapi.Request {
+  payload: (
+    Pick<VolunteerLog, 'id' | 'activity' | 'duration' | 'startedAt' | 'deletedAt'> &
+    { userId: number | string }
+  )[];
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface GetMyVolunteerProjectRequest extends Hapi.Request {
+  params: {
+    projectId: string
+  };
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface PostMyVolunteerProjectRequest extends Hapi.Request {
+  payload: {
+    name: string
+  };
+  pre: {
+    communityBusiness: CommunityBusiness
+  };
+}
+
+export interface PutMyVolunteerProjectRequest extends Hapi.Request {
+  params: {
+    projectId: string
+  };
+  payload: {
+    name: string
+  };
+  pre: {
+    communityBusiness: CommunityBusiness
   };
 }
 
