@@ -5,15 +5,23 @@
  * - /api/v1/api.json
  */
 import get from './get';
+import put from './put';
 import register from './register';
-import adminLogin from './login_admin';
+import adminLogin from './login';
 import logout from './logout';
 import visitors from './visitors';
+import volunteers from './volunteers';
+import password from './password';
+import roles from './roles';
 
 export default [
-  ...register,
   ...get,
+  ...put,
+  ...register,
   ...adminLogin,
   ...logout,
   ...visitors,
+  ...volunteers,
+  ...password,
+  ...roles,
 ];
