@@ -1,9 +1,10 @@
 import validate from './validate';
 import { UserCredentials } from './types';
-import { SessionManager as Session } from './session';
+import { SessionManager as Session, TokenManager as Token } from './session';
 
 export {
   UserCredentials,
   Session,
+  Token,
   validate,
 };
