@@ -1,9 +1,10 @@
 import * as Knex from 'knex';
+import { compare } from 'bcrypt';
 import { omit } from 'ramda';
 import { getConfig } from '../../../config';
 import factory from '../../../tests/utils/factory';
 import { getTrx } from '../../../tests/utils/database';
-import { CbAdmins } from '..';
+import { CbAdmins, DisabilityEnum } from '..';
 
 
 describe('CbAdmin model', () => {
@@ -71,10 +72,11 @@ describe('CbAdmin model', () => {
   describe('Write', () => {
     test('add :: create new record using minimal information', async () => {
       const changeset = await factory.build('cbAdmin');
-
       const admin = await CbAdmins.add(trx, changeset);
+      const passwordCheck = await compare(changeset.password, admin.password);
 
-      expect(admin).toEqual(expect.objectContaining(changeset));
+      expect(admin).toEqual(expect.objectContaining(omit(['password'], changeset)));
+      expect(passwordCheck).toBeTruthy();
     });
 
     test('update :: non-foreign key column', async () => {
@@ -87,7 +89,7 @@ describe('CbAdmin model', () => {
     });
 
     test('update :: foreign key column', async () => {
-      const changeset = { disability: 'no' };
+      const changeset = { disability: DisabilityEnum.NO };
       const admin = await CbAdmins.getOne(trx, { where: { id: 2 } });
 
       const updatedAdmin = await CbAdmins.update(trx, admin, changeset);
@@ -98,7 +100,7 @@ describe('CbAdmin model', () => {
     test('update :: failed update on foreign key column', async () => {
       expect.assertions(1);
 
-      const changeset = { gender: 'non-existent' };
+      const changeset = <any> { gender: 'non-existent' };
       const admin = await CbAdmins.getOne(trx, { where: { id: 2 } });
 
       try {
