@@ -1,6 +1,28 @@
+import Duration from './duration';
+
+
 export { Users } from './user';
 export { Visitors } from './visitor';
+export { Volunteers } from './volunteer';
 export { CbAdmins } from './cb_admin';
 export { Organisations } from './organisation';
 export { CommunityBusinesses } from './community_business';
-export { User, Organisation, CommunityBusiness, LinkedFeedback } from './types';
+export { VolunteerLogs } from './volunteer_log';
+export { Duration };
+export {
+  User,
+  Organisation,
+  CommunityBusiness,
+  FundingBody,
+  LinkedFeedback,
+  VisitActivity,
+  VisitEvent,
+  LinkedVisitEvent,
+  Subscription,
+  VolunteerLog,
+  GenderEnum,
+  DisabilityEnum,
+  EthnicityEnum,
+  ModelQuery,
+  CommonTimestamps,
+} from './types';
