@@ -1,9 +1,9 @@
-enum EmailTemplate {
-  WELCOME_CB = '4010082',
-  WELCOME_VISITOR = '3843402',
-  NEW_VISITOR_FOR_TWINE = '4251043',
-  NEW_VISITOR_FOR_CB = '3853062',
-  PASSWORD_RESET = '4148361',
+export enum EmailTemplate {
+  CB_ADMIN_WELCOME = 4010082,
+  VISITOR_WELCOME = 3843402,
+  TWINE_ADMIN_NEW_VISITOR = 4251043,
+  CB_ADMIN_NEW_VISITOR = 3853062,
+  VISITOR_APP_PASSWORD_RESET = 4148361,
+  ADMIN_APP_PASSWORD_RESET = 8786293,
 }
 
-export { EmailTemplate };
