@@ -1,7 +1,7 @@
 export type Nothing = null;
 export type Maybe<T> = T | Nothing;
 
-export type Map<K extends string | number, V> = {
+export type Map<K extends string | number | symbol, V> = {
   [k in K]: V
 }
 
@@ -9,8 +9,18 @@ export type Dictionary<T> = {
   [key: string]: T
 }
 
+export type ValueOf<T> = T[keyof T]
+
+/*
+ * Recursive definition of deep partial without breaking array types
+ */
 export type DeepPartial<T> = {
-  [P in keyof T]?: DeepPartial<T[P]>
+  [P in keyof T]?:
+    T[P] extends Array<infer U>
+      ? Array<DeepPartial<U>>
+      : T[P] extends ReadonlyArray<infer U>
+        ? ReadonlyArray<DeepPartial<U>>
+        : DeepPartial<T[P]>
 }
 
 export type JsonPrimitives =
